FINAL CODE
Both Switches Functions the way asked In the lab.
For SW1 I used the seg7Display struct variable to reset each digit on the screen including colon.
And For SW2 I changed If - if condtions to If-Else If condition
because I noticed when I run code with if-if condition it will never resum the stopwatch meaning it never checks the 2nd condition!
thats whay if - else if was the better solution.

some commets added in .asm file
Thanks - Vishal